third_party/WebKit/LayoutTests/bluetooth/requestDevice/unnamed-device.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    return setBluetoothFakeAdapter('UnnamedHeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}],
      optionalServices: ['generic_access']}))
    .then(device => {
      assert_equals(device.name, '');
      return device.gatt.connect();
    })
    .then(gattServer => gattServer.getPrimaryService('generic_access'))
    .then(service => service.getCharacteristic('gap.device_name'))
    .then(characteristic => characteristic.readValue())
    .then(value => {
      let decoder = new TextDecoder('utf-8');
      let value_str = decoder.decode(value);
      assert_equals(value_str, '');
    });
}, 'An unnamed device can be obtained by advertised service UUID.');

promise_test(() => {
    return setBluetoothFakeAdapter('UnnamedHeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{name: ['']}],
      optionalServices: ['generic_access']}))
    .then(device => {
      assert_equals(device.name, '');
      return device.gatt.connect();
    })
    .then(gattServer => gattServer.getPrimaryService('generic_access'))
    .then(service => service.getCharacteristic('gap.device_name'))
    .then(characteristic => characteristic.readValue())
    .then(value => {
      let decoder = new TextDecoder('utf-8');
      let value_str = decoder.decode(value);
      assert_equals(value_str, '');
    });
}, 'An unnamed device can be obtained by empty name.');
</script>
